Release Application with Generated Figure

+++ src/main/java/core/basesyntax/Rectangle.java
@@ -21,7 +21,7 @@ public class Rectangle extends Figure {
 
     @Override
     public double getArea() {
-        return (double) (width * height);
+        return (double) width* (double) height;
     }
 
     @Override
